moving camera.py into ./examples/common
