+++ examples/common/camera.py
