+++ examples/markers-2d.py
@@ -29,7 +29,7 @@ sizes[...] = np.linspace(0.05, 12.0,n)**2
 types = core.Marker.star
 markers = visual.Markers(P, types, sizes, None, angles, black, white, 0.5)
 
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("ortho")
 camera.connect(viewport, "motion",  markers.render)
 camera.save("output/markers-2d.png")
