+++ examples/markers-3d.py
@@ -33,7 +33,7 @@ markers = visual.Markers(P,
                          line_colors = white,
                          line_widths = 0.0)
 
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("perspective")
 camera.connect(viewport, "motion",  markers.render)
 camera.save("output/markers-3d.png")
