+++ examples/paths-2d.py
@@ -34,7 +34,7 @@ paths = visual.Paths(P, I.tolist(),
                      line_joins = gsp.core.LineJoin.miter,
                      line_caps = gsp.core.LineCap.round)
 
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("ortho")
 camera.connect(viewport, "motion",  paths.render)
 # camera.save("output/paths-2d.png")
