+++ examples/paths-3d.py
@@ -35,7 +35,7 @@ paths = visual.Paths(P, I,
                      line_joins = gsp.core.LineJoin.round,
                      line_caps = gsp.core.LineCap.round)
 
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("perspective", theta=50, phi=50, zoom = 1.25)
 camera.connect(viewport, "motion",  paths.render)
 # camera.save("output/paths-3d.png")
