+++ examples/pixels-2d.py
@@ -18,7 +18,7 @@ n = 250_000
 P = glm.to_vec3(np.random.uniform(-1, +1, (n,2)))
 pixels = visual.Pixels(P, colors=[0,0,0,1])
 
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("ortho")
 camera.connect(viewport, "motion",  pixels.render)
 camera.save("output/pixels-2d.png")
