+++ examples/pixels-interactive.py
@@ -29,7 +29,7 @@ pixels.render(viewport)
 # -----------------------------------
 
 # Run the camera
-from camera import Camera
+from examples.common.camera import Camera
 camera = Camera("perspective", theta=50, phi=50)
 camera.connect(viewport, "motion",  pixels.render)
 camera.run()
